fix: added eodashCompareCollections state to adapt compare map + moved EodashMap logic to composables
