+++ core/client/store/stac.js
@@ -5,7 +5,7 @@ import { useAbsoluteUrl, useCompareAbsoluteUrl } from "@/composables/index";
 import { eodashKey } from "@/utils/keys";
 import { indicator } from "@/store/States";
 import { extractCollectionUrls } from "@/utils/helpers";
-import { eodashCollections } from "@/utils/states";
+import { eodashCollections, eodashCompareCollections } from "@/utils/states";
 import { EodashCollection } from "@/utils/eodashSTAC";
 
 export const useSTAcStore = defineStore("stac", () => {
@@ -84,28 +84,27 @@ export const useSTAcStore = defineStore("stac", () => {
     await axios
       .get(absoluteUrl.value)
       .then(async (resp) => {
-        selectedStac.value = resp.data;
-        indicator.value = selectedStac.value?.id ?? "";
-
         // init eodash collections
         const collectionUrls = extractCollectionUrls(
-          selectedStac.value,
+          resp.data,
           absoluteUrl.value,
         );
 
-        // empty array from old collections
-        eodashCollections.length = 0;
+        await Promise.all(
+          collectionUrls.map((cu) => {
+            const ec = new EodashCollection(cu);
+            ec.fetchCollection();
+            return ec;
+          }),
+        ).then((collections) => {
+          // empty array from old collections
+          eodashCollections.length = 0;
+          // update eodashCollections
+          eodashCollections.push(...collections);
+        });
 
-        // update eodashCollections
-        eodashCollections.push(
-          ...(await Promise.all(
-            collectionUrls.map((cu) => {
-              const ec = new EodashCollection(cu);
-              ec.fetchCollection();
-              return ec;
-            }),
-          )),
-        );
+        selectedStac.value = resp.data;
+        indicator.value = selectedStac.value?.id ?? "";
       })
       .catch((err) => {
         throw new Error("error loading the selected STAC", err);
@@ -125,7 +124,26 @@ export const useSTAcStore = defineStore("stac", () => {
 
     await axios
       .get(absoluteUrl.value)
-      .then((resp) => {
+      .then(async (resp) => {
+        // init eodash collections
+        const collectionUrls = extractCollectionUrls(
+          resp.data,
+          absoluteUrl.value,
+        );
+
+        await Promise.all(
+          collectionUrls.map((cu) => {
+            const ec = new EodashCollection(cu);
+            ec.fetchCollection();
+            return ec;
+          }),
+        ).then((collections) => {
+          // empty array from old collections
+          eodashCompareCollections.length = 0;
+          // update eodashCompareCollections
+          eodashCompareCollections.push(...collections);
+        });
+
         selectedCompareStac.value = resp.data;
       })
       .catch((err) => {
